serial send and serial receive both working

+++ src/avionics_costco/include/packet_definition.hpp
@@ -97,7 +97,7 @@ struct MassPacket {
 
 struct ServoRequest {
     uint16_t id;
-    float increment;
+    uint8_t increment;
     bool zero_in;
 };
